+++ CHANGES.markdown
@@ -3,6 +3,7 @@
   - Remove `BuildSpecs` constructor from `SpecTree`, having `runIO` this is not
     really needed
   - Add `mapSpecTree` to `Test.Hspec.Core`
+  - Add optional source location to spec items
   - Add `afterAll` (see #188)
   - Omit empty `describe` groups from report
   - Do not print empty lines before/after spec groups